feat: 과목 별 학점 조회 기능 추가

+++ src/main/java/Iconicthon/moadream/domain/Credit/dto/response/CreditResponse.java
@@ -21,7 +21,7 @@ public record CreditResponse (
         Integer bsm,
 
         @JsonProperty("average_grade")
-        Integer averageGrade,
+        Float averageGrade,
 
         @JsonProperty("english_count")
         long englishCount
